Add tests for dashboard server (#39)
This change uses dependency injection to pass the GitHub client into the dashboard server to allow for testing. The test then re-uses the recording/replay server for testing against the GitHub API. The test uses a test account `project-health1` so that tests are real, but unchanging. To re-record the API responses, any GitHub token must be used with the test user still fixed.