Fix globstar patterns in package.json.
We no longer rely on the shell to do ** globstar expansion in an NPM script,
which is important, since it won't (at least not on macOS by default, or
windows).

It seems like we were also not linting all the files, even on Travis, so
this also adds some lint fixes.