Sort incoming pull requests (#184)
This change sorts the list of incoming pull requests that are displayed. They are ordered first by whether or not the status is actionable. Next they are sorted by the latest event time, which excludes your review event if any.

This allows new activity to bubble to the top, while maintaining an actionable view.