Handle cases where there has been no submitted review
Fixes #517.
In some cases, a PR will appear in GitHubs response under reviewed-by: even though no review has been submitted yet. This results in the PR being incorrectly handled. This fixes it by not handling any PR which has review when we did expect a review. This will allow it to fall through and be handled by other logic, such as what review-requests are still outstanding.